Finance/requirements.txt
Flask==2.3.3
fastapi==0.111.0  # Newer version compatible with pydantic v2
pydantic==2.8.2   # Latest pydantic version as of now
mysql-connector-python==8.1.0
uvicorn==0.23.2
jinja2==3.1.2
